Differential spell checking
This script does not attempt to find all misspellings, but rather show
when new words not previously known are introduced. This means it's
useful in a continuous integration pipeline where we don't necessarily
want each contributor to have to be responsible for the bad spelling of
whoever came before them, but rather at least not make things worse.
The existing spelling can then be incrementally cleaned up.

This approach also means we don't need to maintain a local dictionary to
avoid false-positives, because existing "misspelled words" ("AppVM",
"marmarek", etc.) will be ignored.

In addition, effort is made to only check prose, not code. All the
actual parsing logic is handled by pandoc, and we just filter and
transform a JSON AST that it produces.

We also operate on the fully built HTML site rather than the source
markdown, meaning we see all spellings as they exist post-templating,
including all values from _data/*.yaml and such already filled in.

This was written primarily for use in a CI hook, but is also useful
locally. It can be used in non-differential mode to show all existing
"misspellings" (with lots of false positives), or in differential mode
if you keep around an old copy of the built site. This only operates on
built copies of the site, so it is up to the caller to keep around an
old copy of the built site to compare against.

A previous version of this script which attempted to check out and build
the correct sites from git for comparison can be found here:
- https://github.com/QubesOS/qubesos.github.io/pull/110
It was determined that this was not the right way to approach the
problem.

Introduced dependencies are:
- hunspell (preferred) or aspell
- pandoc
- jq

Fixes: https://github.com/QubesOS/qubes-issues/issues/2725